requirements.txt
alabaster==0.7.12
alembic==1.6.5
anyio==3.2.1
apache-airflow==2.1.0
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
argcomplete==1.12.3
arrow==1.2.2
asttokens==2.0.5
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.12b0
bs4==0.0.1
cattrs==1.7.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cssselect==1.1.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
entrypoints==0.3
executing==0.8.2
Flask==1.1.4
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
graphviz==0.16
gunicorn==20.1.0
h11==0.12.0
html2json==0.2.4.1
httpcore==0.13.6
httpx==0.18.2
idna==2.10
imagesize==1.3.0
importlib-metadata==4.10.1
importlib-resources==1.5.0
inflection==0.5.1
ipykernel==6.7.0
ipython==8.0.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.1
jellyfish==0.8.2
Jinja2==3.0.3
jinja2-time==0.2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.20.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.2.4
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prison==0.1.3
prompt-toolkit==3.0.26
psycopg2==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9
Pygments==2.11.2
pyparsing==3.0.7
pyquery==1.4.3
pyrsistent==0.18.0
pyspark==3.1.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-editor==1.0.4
python-Levenshtein==0.12.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
pyzmq==20.0.0
requests==2.25.1
rfc3986==1.5.0
rich==10.4.0
selenium==3.141.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.2.1
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
stack-data==0.1.4
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
unicodecsv==0.14.1
Unidecode==1.3.2
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3
zipp==3.7.0
